update breedSlide to show breedSize; saved data when registering; update coreData for weight from float to Int32; changed dogModel to have relationship to weight

+++ doggo/Helper/CoreDataHelper.swift
@@ -54,7 +54,7 @@ class CoreDataHelper {
         dog.size = model.breedSize
         dog.gender = model.gender
         
-        dogWeight.amount = model.weight.amount
+        dogWeight.amount = Int32(model.weight.amount)
         dogWeight.date = model.weight.date
         dog.addToWeight(dogWeight)
